% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZUMensemble.R
\name{ensemble}
\alias{ensemble}
\title{Creates an ensemble model based on any classification algorithm}
\usage{
ensemble(algorithm, formula, dataset, mcount = 100, pattr = 0.6, ...)
}
\arguments{
\item{algorithm}{a classification algorithm}

\item{formula}{formula defining what attribute to classify}

\item{dataset}{dataframe containing training data}

\item{mcount}{number of models in the ensemble model}

\item{pattr}{percent of all attributes to be sampled for one model}

\item{...}{named arguments passed to a single model}
}
\description{
Creates an ensemble model based on any classification algorithm
}
