% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZUMensemble.R
\docType{package}
\name{ZUMensemble-package}
\alias{ZUMensemble}
\alias{ZUMensemble-package}
\title{ZUMensemble: Ensemble classification model}
\description{
Builds an ensemble model using any classification algorithm passed as an argument.
}
\details{
Build an ensemble classification model with [ensemble()].

Predict the results with [predict()].
}
\keyword{internal}
